# specs/001-paracite-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ParaCite API
  version: 0.1.0
  description: Citas jurídicas de precisión. Solo matches por encima del umbral. Corpus fixture ES.
servers:
  - url: http://127.0.0.1:18741
paths:
  /health:
    get:
      security: []
      responses:
        "200":
          description: OK
  /v1/me:
    get:
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Perfil y cuota
        "401":
          description: Unauthorized
  /v1/cite:
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CiteRequest"
      responses:
        "200":
          description: Matches de alta confianza o vacío
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CiteResponse"
        "401":
          description: Unauthorized
        "422":
          description: Validation error
        "429":
          description: Rate limited
  /v1/ingest:
    post:
      security:
        - bearerAuth: []
      responses:
        "202":
          description: Enqueued when ENABLE_INGEST
        "501":
          description: Enterprise stub
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    CiteRequest:
      type: object
      required: [tesis]
      properties:
        tesis:
          type: string
          minLength: 1
        jurisdiccion:
          type: string
          default: ES
        tipos:
          type: array
          items:
            type: string
            enum: [norma, doctrina, sentencia_fixture]
        umbral_confianza:
          type: number
          default: 0.87
          minimum: 0
          maximum: 1
        max_resultados:
          type: integer
          default: 3
          minimum: 1
          maximum: 20
        incluir_texto_completo:
          type: boolean
          default: false
    CiteMatch:
      type: object
      required: [id, tipo, confianza, cita_formal, parrafo_exacto, enlace_profundo, metadatos]
      properties:
        id: { type: string }
        tipo: { type: string }
        confianza: { type: number }
        cita_formal: { type: string }
        parrafo_exacto: { type: string }
        enlace_profundo: { type: string }
        metadatos: { type: object }
    CiteResponse:
      type: object
      required: [request_id, tesis, matches, sin_match_alta_confianza, tiempo_procesamiento_ms]
      properties:
        request_id: { type: string }
        tesis: { type: string }
        matches:
          type: array
          items:
            $ref: "#/components/schemas/CiteMatch"
        sin_match_alta_confianza: { type: boolean }
        tiempo_procesamiento_ms: { type: integer }